add report page
